requirements.txt
flask
torch
transformers
python-dotenv
